Bug 1195335 - Split wiki views into separate modules.
This includes some cleanups and code deletions.
