--- kuma/wiki/templates/wiki/confirm_revision_delete.html
@@ -1,29 +1,0 @@
-{% extends "wiki/base.html" %}
-{% set title = _('Delete Revision | {document}')|f(document=document.title) %}
-{% block title %}{{ page_title(title) }}{% endblock %}
-
-{% block content %}
-
-      <article id="delete-revision">
-
-        <h1>{{ _('Are you sure you want to delete this revision?') }}</h1>
-
-        <label><strong>{{ _('Creator') }}</strong></label>
-        <div>{{ revision.creator }}</div>
-        <label><strong>{{ _('Date') }}</strong></label>
-        <div>{{ datetimeformat(revision.created, format='longdatetime') }}</div>
-        <label><strong>{{ _('Content') }}</strong></label>
-        <pre>{{ revision.content }}</pre>
-        <form action="" method="post">
-          {{ csrf() }}
-          <p>
-            {{ _('You are about to permanently delete this revision. <strong>This cannot be undone!</strong> Are you sure you want to continue?')|safe }}
-          </p>
-          <div class="submit">
-            <input type="submit" value="{{ _('Delete') }}" />
-            <a href="{{ url('wiki.document_revisions', document.slug) }}">{{ _('Cancel') }}</a>
-          </div>
-        </form>
-      </article>
-
-{% endblock %}
