+++ kuma/wiki/urls.py
@@ -10,146 +10,144 @@ from .constants import DOCUMENT_PATH_RE
 # These patterns inherit (?P<document_path>[^\$]+).
 document_patterns = [
     url(r'^$',
-        views.document,
+        views.document.document,
         name='wiki.document'),
     url(r'^\$revision/(?P<revision_id>\d+)$',
-        views.revision,
+        views.revision.revision,
         name='wiki.revision'),
     url(r'^\$history$',
-        views.document_revisions,
+        views.list.revisions,
         name='wiki.document_revisions'),
     url(r'^\$edit$',
-        views.edit_document,
+        views.edit.edit,
         name='wiki.edit_document'),
     url(r'^\$edit/(?P<revision_id>\d+)$',
-        views.edit_document,
+        views.edit.edit,
         name='wiki.new_revision_based_on'),
     url(r'^\$compare$',
-        views.compare_revisions,
+        views.revision.compare,
         name='wiki.compare_revisions'),
     url(r'^\$children$',
-        views.get_children,
-        name='wiki.get_children'),
+        views.document.children,
+        name='wiki.children'),
     url(r'^\$flag',
-        views.flag,
+        views.misc.flag,
         name='wiki.flag_document'),
     url(r'^\$translate$',
-        views.translate,
+        views.translate.translate,
         name='wiki.translate'),
     url(r'^\$locales$',
-        views.select_locale,
+        views.translate.select_locale,
         name='wiki.select_locale'),
     url(r'^\$json$',
-        views.json_view,
+        views.document.as_json,
         name='wiki.json_slug'),
     url(r'^\$styles$',
-        views.styles_view,
+        views.document.styles,
         name='wiki.styles'),
     url(r'^\$toc$',
-        views.toc_view,
+        views.document.toc,
         name='wiki.toc'),
     url(r'^\$move$',
-        views.move,
+        views.document.move,
         name='wiki.move'),
     url(r'^\$quick-review$',
-        views.quick_review,
+        views.revision.quick_review,
         name='wiki.quick_review'),
     url(r'^\$samples/(?P<sample_id>.+)/files/(?P<attachment_id>\d+)/(?P<filename>.+)$',
-        views.raw_code_sample_file,
+        views.code.raw_code_sample_file,
         name='wiki.raw_code_sample_file'),
     url(r'^\$samples/(?P<sample_id>.+)$',
-        views.code_sample,
+        views.code.code_sample,
         name='wiki.code_sample'),
     url(r'^\$revert/(?P<revision_id>\d+)$',
-        views.revert_document,
+        views.delete.revert_document,
         name='wiki.revert_document'),
     url(r'^\$repair_breadcrumbs$',
-        views.repair_breadcrumbs,
+        views.document.repair_breadcrumbs,
         name='wiki.repair_breadcrumbs'),
     url(r'^\$delete$',
-        views.delete_document,
+        views.delete.delete_document,
         name='wiki.delete_document'),
     url(r'^\$restore$',
-        views.restore_document,
+        views.delete.restore_document,
         name='wiki.restore_document'),
     url(r'^\$purge$',
-        views.purge_document,
+        views.delete.purge_document,
         name='wiki.purge_document'),
 
     # Un/Subscribe to document edit notifications.
     url(r'^\$subscribe$',
-        views.subscribe_document,
-        name='wiki.subscribe_document'),
+        views.document.subscribe,
+        name='wiki.subscribe'),
 
     # Vote helpful/not helpful
     url(r'^\$vote',
-        views.helpful_vote,
+        views.misc.helpful_vote,
         name="wiki.document_vote"),
 ]
 
 urlpatterns = [
-    # Un/Subscribe to locale 'ready for review' notifications.
     url(r'^/ckeditor_config.js$',
-        views.ckeditor_config,
+        views.misc.ckeditor_config,
         name='wiki.ckeditor_config'),
 
     # internals
     url(r'^.json$',
-        views.json_view,
+        views.document.as_json,
         name='wiki.json'),
     url(r'^/preview-wiki-content$',
-        views.preview_revision,
+        views.revision.preview,
         name='wiki.preview'),
     url(r'^/move-requested$',
-        TemplateView.as_view(
-            template_name='wiki/move_requested.html'),
+        TemplateView.as_view(template_name='wiki/move_requested.html'),
         name='wiki.move_requested'),
     url(r'^/get-documents$',
-        views.autosuggest_documents,
+        views.misc.autosuggest_documents,
         name='wiki.autosuggest_documents'),
     url(r'^/load/$',
-        views.load_documents,
+        views.misc.load_documents,
         name='wiki.load_documents'),
 
     # Special pages
     url(r'^/templates$',
-        views.list_templates,
+        views.list.templates,
         name='wiki.list_templates'),
     url(r'^/tags$',
-        views.list_tags,
+        views.list.tags,
         name='wiki.list_tags'),
     url(r'^/tag/(?P<tag>.+)$',
-        views.list_documents,
+        views.list.documents,
         name='wiki.tag'),
     url(r'^/new$',
-        views.new_document,
+        views.create.new_document,
         name='wiki.new_document'),
     url(r'^/all$',
-        views.list_documents,
+        views.list.documents,
         name='wiki.all_documents'),
     url(r'^/with-errors$',
-        views.list_documents_with_errors,
+        views.list.with_errors,
         name='wiki.errors'),
     url(r'^/without-parent$',
-        views.list_documents_without_parent,
+        views.list.without_parent,
         name='wiki.without_parent'),
     url(r'^/top-level$',
-        views.list_top_level_documents,
+        views.list.top_level,
         name='wiki.top_level'),
     url(r'^/needs-review/(?P<tag>[^/]+)$',
-        views.list_documents_for_review,
+        views.list.needs_review,
         name='wiki.list_review_tag'),
     url(r'^/needs-review/?',
-        views.list_documents_for_review,
+        views.list.needs_review,
         name='wiki.list_review'),
     url(r'^/localization-tag/(?P<tag>[^/]+)$',
-        views.list_documents_with_localization_tag,
+        views.list.with_localization_tag,
         name='wiki.list_with_localization_tag'),
     url(r'^/localization-tag/?',
-        views.list_documents_with_localization_tag,
+        views.list.with_localization_tag,
         name='wiki.list_with_localization_tags'),
     url(r'^/category/(?P<category>\d+)$',
-        views.list_documents,
+        views.list.documents,
         name='wiki.category'),
 
     # Feeds
