+++ kuma/wiki/views/__init__.py
@@ -1,0 +1,6 @@
+import logging
+
+from . import (code, create, delete, document, edit, legacy, list,  # noqa
+               misc, revision, translate)  # noqa
+
+log = logging.getLogger('kuma.wiki.views')
